:bug: Fixes refresh bug
