+++ lib/exams/blocs/exam_event.dart
@@ -1,30 +1,30 @@
-part of 'exam_bloc.dart';
+// part of 'exam_bloc.dart';
 
-abstract class ExamEvent extends Equatable {
-  @override
-  List<Object> get props => [];
-}
+// abstract class ExamEvent extends Equatable {
+//   @override
+//   List<Object> get props => [];
+// }
 
-class SaveExam extends ExamEvent {
-  final cardExamInfo;
-  final examDetails;
-  final examFile;
+// class SaveExam extends ExamEvent {
+//   final cardExamInfo;
+//   final examDetails;
+//   final examFile;
 
-  SaveExam(
-      {@required this.cardExamInfo,
-      @required this.examDetails,
-      @required this.examFile});
+//   SaveExam(
+//       {@required this.cardExamInfo,
+//       @required this.examDetails,
+//       @required this.examFile});
 
-  CardExamInfo get getCardExamInfo => this.cardExamInfo;
-  ExamDetails get getExamDetails => this.examDetails;
-  File get getExamFile => this.examFile;
-}
+//   CardExamInfo get getCardExamInfo => this.cardExamInfo;
+//   ExamDetails get getExamDetails => this.examDetails;
+//   File get getExamFile => this.examFile;
+// }
 
-// TODO: Alterar evento se necessário
-class GetExams extends ExamEvent {}
+// // TODO: Alterar evento se necessário
+// class GetExams extends ExamEvent {}
 
-class DecriptExam extends ExamEvent {
-  final filePath;
+// class DecriptExam extends ExamEvent {
+//   final filePath;
 
-  DecriptExam({@required this.filePath});
-}
+//   DecriptExam({@required this.filePath});
+// }
