+++ lib/exams/blocs/exam_state.dart
@@ -1,38 +1,38 @@
-part of 'exam_bloc.dart';
+// part of 'exam_bloc.dart';
 
-abstract class ExamState extends Equatable {
-  @override
-  List<Object> get props => [];
-}
+// abstract class ExamState extends Equatable {
+//   @override
+//   List<Object> get props => [];
+// }
 
-class ExamInitial extends ExamState {}
+// class ExamInitial extends ExamState {}
 
-class ExamProcessing extends ExamState {}
+// class ExamProcessing extends ExamState {}
 
-class GetExamsSuccess extends ExamState {
-  final cardExamInfo;
-  final examDetails;
-  final filePath;
+// class GetExamsSuccess extends ExamState {
+//   final cardExamInfo;
+//   final examDetails;
+//   final filePath;
 
-  GetExamsSuccess(
-      {@required this.cardExamInfo,
-      @required this.examDetails,
-      @required this.filePath});
+//   GetExamsSuccess(
+//       {@required this.cardExamInfo,
+//       @required this.examDetails,
+//       @required this.filePath});
 
-  CardExamInfo get getCardExamInfo => this.cardExamInfo;
-  ExamDetails get getExamDetails => this.examDetails;
-  String get getFilePath => this.filePath;
-}
+//   CardExamInfo get getCardExamInfo => this.cardExamInfo;
+//   ExamDetails get getExamDetails => this.examDetails;
+//   String get getFilePath => this.filePath;
+// }
 
-class ExamProcessingSuccess extends ExamState {
-  File encriptedFile;
-  ExamProcessingSuccess({@required this.encriptedFile});
-}
+// class ExamProcessingSuccess extends ExamState {
+//   File encriptedFile;
+//   ExamProcessingSuccess({@required this.encriptedFile});
+// }
 
-class DecriptExamSuccess extends ExamState {
-  final decriptedBytes;
+// class DecriptExamSuccess extends ExamState {
+//   final decriptedBytes;
 
-  DecriptExamSuccess({@required this.decriptedBytes});
-}
+//   DecriptExamSuccess({@required this.decriptedBytes});
+// }
 
-class ExamProcessingFail extends ExamState {}
+// class ExamProcessingFail extends ExamState {}
