+++ lib/main.dart
@@ -13,6 +13,7 @@ import 'package:tcc_projeto_app/login/blocs/login_bloc.dart';
 import 'package:tcc_projeto_app/login/blocs/signup_bloc.dart';
 import 'package:tcc_projeto_app/login/repositories/user_repository.dart';
 import 'package:tcc_projeto_app/login/screens/login_screen.dart';
+import 'package:tcc_projeto_app/med_record/blocs/med_record_bloc.dart';
 import 'package:tcc_projeto_app/med_record/repositories/med_record_repository.dart';
 import 'package:tcc_projeto_app/pacient/repositories/pacient_repository.dart';
 import 'package:tcc_projeto_app/route_generator.dart';
@@ -72,12 +73,12 @@ class _MyAppState extends State<MyApp> {
             authenticationBloc: this.authenticationBloc,
           ),
         ),
-        BlocProvider<ExamBloc>(
-          create: (context) => ExamBloc(
-            examRepository:
-                Injector.appInstance.getDependency<ExamRepository>(),
-          ),
-        )
+        BlocProvider<MedRecordBloc>(
+            create: (context) => MedRecordBloc(
+                medRecordRepository:
+                    Injector.appInstance.getDependency<MedRecordRepository>(),
+                examRepository:
+                    Injector.appInstance.getDependency<ExamRepository>()))
       ],
       child: MaterialApp(
         title: "Projeto tcc",
