+++ lib/med_record/blocs/med_record_bloc.dart
@@ -1,8 +1,15 @@
 import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
+import 'dart:math';
 
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:tcc_projeto_app/exams/models/card_exam_info.dart';
+import 'package:tcc_projeto_app/exams/models/exam_details.dart';
+import 'package:tcc_projeto_app/exams/repositories/exam_repository.dart';
 import 'package:tcc_projeto_app/med_record/repositories/med_record_repository.dart';
 
 part 'med_record_event.dart';
@@ -10,8 +17,11 @@ part 'med_record_state.dart';
 
 class MedRecordBloc extends Bloc<MedRecordEvent, MedRecordState> {
   MedRecordRepository medRecordRepository;
+  ExamRepository examRepository;
 
-  MedRecordBloc({@required this.medRecordRepository}) : super(null);
+  MedRecordBloc(
+      {@required this.medRecordRepository, @required this.examRepository})
+      : super(null);
 
   @override
   MedRecordState get initialState => MedRecordInicialState();
@@ -42,6 +52,56 @@ class MedRecordBloc extends Bloc<MedRecordEvent, MedRecordState> {
       try {} catch (error) {}
     } else if (event is MedRecordDeleteButtonPressed) {
       try {} catch (error) {}
+    }
+    if (event is SaveExam) {
+      try {
+        yield ExamProcessing();
+
+        var examBytes = await event.getExamFile.readAsBytes();
+
+        var encoded = base64.encode(examBytes);
+
+        String dir = (await getApplicationDocumentsDirectory()).path;
+        var randomFileName = Random.secure().nextInt(10000);
+        var filePath = "$dir/$randomFileName.txt";
+        var encriptedFile = File("$dir/$randomFileName.txt");
+        await encriptedFile.writeAsString(encoded);
+
+        await this
+            .examRepository
+            .saveExam(event.getCardExamInfo, event.getExamDetails, filePath);
+
+        yield ExamProcessingSuccess(encriptedFile: encriptedFile);
+      } catch (error) {
+        yield ExamProcessingFail();
+      }
+    } else if (event is GetExams) {
+      try {
+        yield ExamProcessing();
+
+        var examInfoList = await this.examRepository.getExam();
+
+        yield GetExamsSuccess(
+            cardExamInfo: examInfoList.first,
+            examDetails: examInfoList[1],
+            filePath: examInfoList.last);
+      } catch (error) {
+        yield ExamProcessingFail();
+      }
+    } else if (event is DecriptExam) {
+      try {
+        yield ExamProcessing();
+
+        var filePath = event.filePath;
+
+        var file = File(filePath);
+        var bytes = await file.readAsString();
+        var decriptedBytes = base64.decode(bytes);
+
+        yield DecriptExamSuccess(decriptedBytes: decriptedBytes);
+      } catch (error) {
+        yield ExamProcessingFail();
+      }
     }
   }
 }
