+++ lib/med_record/blocs/med_record_event.dart
@@ -10,24 +10,60 @@ class MedRecordCreateButtonPressed extends MedRecordEvent {
       );
 
   @override
-  List<Object> get props => throw UnimplementedError();
+  List<Object> get props => [];
 }
 
 class MedRecordLoad extends MedRecordEvent {
   @override
-  List<Object> get props => throw UnimplementedError();
+  List<Object> get props => [];
 }
 
 class MedRecordEditButtonPressed extends MedRecordEvent {
   MedRecordEditButtonPressed();
 
   @override
-  List<Object> get props => throw UnimplementedError();
+  List<Object> get props => [];
 }
 
 class MedRecordDeleteButtonPressed extends MedRecordEvent {
   MedRecordDeleteButtonPressed();
 
   @override
-  List<Object> get props => throw UnimplementedError();
+  List<Object> get props => [];
+}
+
+class SaveExam extends MedRecordEvent {
+  final cardExamInfo;
+  final examDetails;
+  final examFile;
+
+  SaveExam(
+      {@required this.cardExamInfo,
+      @required this.examDetails,
+      @required this.examFile});
+
+  CardExamInfo get getCardExamInfo => this.cardExamInfo;
+  ExamDetails get getExamDetails => this.examDetails;
+  File get getExamFile => this.examFile;
+
+  @override
+  // TODO: implement props
+  List<Object> get props => [];
+}
+
+// TODO: Alterar evento se necessário
+class GetExams extends MedRecordEvent {
+  @override
+  // TODO: implement props
+  List<Object> get props => [];
+}
+
+class DecriptExam extends MedRecordEvent {
+  final filePath;
+
+  DecriptExam({@required this.filePath});
+
+  @override
+  // TODO: implement props
+  List<Object> get props => [];
 }
